Create DAO and Entities

+++ src/main/kotlin/org/example/gamebusters/model/InfoGameJson.kt
@@ -1,8 +1,10 @@
 package org.example.gamebusters.model
 
+import java.math.BigDecimal
+
 data class InfoGameJson(
     val titulo: String,
     val capa: String,
-    val preco: Double,
+    val preco: BigDecimal,
     val descricao: String
 ){}
